feat: ✨ Character Generator
This is a working implementation of the character generator. Courtesy of @maxtar.

> Note! Static files, inlcuding system.json have been modified.

- The character generator can be accessed from the character sheet via the leaf icon.

!!Bug At the moment it only creates a new character. This will need to be fixed.
