+++ gulpfile.js
@@ -21,7 +21,7 @@ const stylesDirectory = `${sourceDirectory}/styles`;
 const stylesExtension = "scss";
 const sourceFileExtension = "js";
 const srcFiles = ["lang", "templates"];
-const staticFiles = ["assets", "fonts", "system.json", "template.json"];
+const staticFiles = ["assets", "fonts", "scripts", "system.json", "template.json"];
 const getDownloadURL = (version) =>
 	`https://github.com/fvtt-fria-ligan/forbidden-lands-foundry-vtt/releases/download/${version}/${version}.zip`;
 const repoPathing = (relativeSourcePath = ".", sourcemapPath = ".") => {
