--- src/module/character-generator/lang/en.json
@@ -1,18 +1,0 @@
-{
-    "FLCG.TITLE": "Character Generator",
-    "FLCG.RANDOMIZE": "Randomize All",
-    "FLCG.CREATE": "Create Actor",
-    "FLCG.KIN": "Kin",
-    "FLCG.AGE": "Age",
-    "FLCG.AGE_PENALTY_MESSAGE": "Attribute penalty from age will be applied randomly",
-    "FLCG.CHILDHOOD": "Childhood",
-    "FLCG.PROFESSION": "Profession",
-    "FLCG.PATH": "Path",
-    "FLCG.EVENT": "Formative Event",
-    "FLCG.SKILLS": "Skills",
-    "FLCG.ATTRIBUTES": "Attributes",
-    "FLCG.TALENT": "Talent",
-
-    "FLCG.SETTINGS.DATASET_DIR": "Dataset Directory",
-    "FLCG.SETTINGS.DATASET_DIR_HINT": "Path to a directory with dataset.json file. You can make your own dataset to customize generator's output."
-}
