--- src/module/character-generator/templates/character-generator.html
@@ -1,97 +1,0 @@
-<form class="{{cssClass}}" autocomplete="off">
-    <div class="chargen flex column">
-        <div class="flex row chargen-section border">
-            <button class="chargen-randomize-all" type="button">{{localize "FLCG.RANDOMIZE"}}</button>
-            <button class="chargen-create-actor" type="button">{{localize "FLCG.CREATE"}}</button>
-        </div>
-
-        <div class="flex row chargen-section border">
-            <button class="chargen-roll chargen-roll-kin" type="button"></button>
-            <div class="chargen-section-content">
-                {{localize "FLCG.KIN"}}
-                <select class="chargen-select-kin">
-                    {{#select character.kin}}
-                    {{#each dataset.kin as |item|}}
-                        <option value="{{item.key}}">{{localize item.name}}</option>
-                    {{/each}}
-                    {{/select}}
-                </select>
-            </div>
-            <button class="chargen-roll chargen-roll-age" type="button"></button>
-            <div class="chargen-section-content">
-                {{localize "FLCG.AGE"}}
-                <input class="chargen-age-input" type="number" maxlength="3" value="{{character.age.ageNumber}}"/>
-                <span class="chargen-age">{{localize character.age.ageString}}</span>
-            </div>
-            <div class="chargen-section-content">
-                {{localize "FLCG.AGE_PENALTY_MESSAGE"}}
-            </div>
-        </div>
-
-        <div class="flex row chargen-section border">
-            <button class="chargen-roll chargen-roll-childhood" type="button"></button>
-            <div class="chargen-section-content">
-                {{localize "FLCG.CHILDHOOD"}}
-                <select class="chargen-select-childhood">
-                    {{#select character.childhood.key}}
-                    {{#each dataset.childhood as |item|}}
-                        <option value="{{item.key}}">{{localize item.name}}</option>
-                    {{/each}}
-                    {{/select}}
-                </select>
-                <div class="chargen-description">{{character.childhood.description}}</div>
-            </div>
-            {{> modules/forbidden-lands-character-generator/templates/partial/list.html title="FLCG.ATTRIBUTES" list=character.childhood.attributes }}
-            {{> modules/forbidden-lands-character-generator/templates/partial/list.html title="FLCG.SKILLS" list=character.childhood.skills }}
-        </div>
-
-        <div class="flex row chargen-section border">
-            <button class="chargen-roll chargen-roll-profession" type="button"></button>
-            <div class="chargen-section-content">
-                {{localize "FLCG.PROFESSION"}}
-                <select class="chargen-select-profession">
-                    {{#select character.profession}}
-                    {{#each dataset.profession as |item|}}
-                        <option value="{{item.key}}">{{localize item.name}}</option>
-                    {{/each}}
-                    {{/select}}
-                </select>
-            </div>
-
-            <button class="chargen-roll chargen-roll-path" type="button"></button>
-            <div class="chargen-section-content">
-                {{localize "FLCG.PATH"}}
-                <select class="chargen-select-path">
-                    {{#select character.path}}
-                    {{#each dataset.paths as |item key|}}
-                        <option value="{{key}}">{{localize item}}</option>
-                    {{/each}}
-                    {{/select}}
-                </select>
-            </div>
-        </div>
-        
-        {{#each character.formativeEvents as |formativeEvent key|}}
-            <div class="flex row chargen-section border">
-                <button class="chargen-roll chargen-roll-event" data-key="{{key}}" type="button"></button>
-                <div class="chargen-section-content">
-                    {{localize "FLCG.EVENT"}}
-                    <select class="chargen-select-event" data-key="{{key}}">
-                        {{#select formativeEvent.key}}
-                        {{#each ../dataset.formativeEvents as |item|}}
-                            <option value="{{item.key}}">{{localize item.name}}</option>
-                        {{/each}}
-                        {{/select}}
-                    </select>
-                    <div class="chargen-description chargen-event">{{formativeEvent.description}}</div>
-                    <div class="chargen-items chargen-event">{{formativeEvent.items}}</div>
-                </div>
-                {{> modules/forbidden-lands-character-generator/templates/partial/list.html title="FLCG.SKILLS" list=formativeEvent.skills }}
-                <div class="chargen-list">
-                    <div class="chargen-caption chargen-talent">{{localize "FLCG.TALENT"}}</div>
-                    <div class="chargen-description chargen-talent">{{formativeEvent.talent}}</div>
-                </div>
-            </div>
-        {{/each}}
-    </div>
-</form>
