--- src/module/character-generator/templates/partial/list.html
@@ -1,9 +1,0 @@
-
-<div class="chargen-list">
-    <div class="chargen-list-caption">{{localize title}}</div>
-    <ul>
-    {{#each list as |value key|}}
-        <li>{{key}}: {{value}}</li>
-    {{/each}}
-    </ul>
-</div>
