+++ src/module/hooks/handlebars.js
@@ -34,6 +34,7 @@ function preloadHandlebarsTemplates() {
 		"systems/forbidden-lands/templates/partial/travel-action.hbs",
 		"systems/forbidden-lands/templates/partial/party-member.hbs",
 		"systems/forbidden-lands/templates/party.hbs",
+		"systems/forbidden-lands/templates/partial/list.hbs",
 	];
 	return loadTemplates(templatePaths);
 }
@@ -47,6 +48,11 @@ function registerHandlebarsHelpers() {
 			acc += block.fn(this);
 		}
 		return acc;
+	});
+	Handlebars.registerHelper("flps_enrich", function (content) {
+		// Enrich the content
+		content = TextEditor.enrichHTML(content, { entities: true });
+		return new Handlebars.SafeString(content);
 	});
 	Handlebars.registerHelper("flps_capitalize", function (value) {
 		return typeof value === "string" && value.length > 0 ? value[0].toUpperCase() + value.slice(1) : value;
