+++ src/styles/components/_character-generator.scss
