+++ src/styles/components/_index.scss
@@ -1,1 +1,2 @@
 @forward "chat";
+@forward "character-generator";
