+++ src/templates/partial/list.hbs
@@ -1,0 +1,8 @@
+<div class="chargen-list">
+	<div class="chargen-list-caption">{{localize title}}</div>
+	<ul>
+		{{#each list as |value key|}}
+		<li>{{key}}: {{value}}</li>
+		{{/each}}
+	</ul>
+</div>
